refactor: move SpawnGate to freshell-freshagent (neutral home below both create doors)
Behavior-preserving: freshell-ws re-exports the moved module; tracing
targets and gate semantics byte-identical; from_config inlined at its
two call sites. Groundwork for gating the REST create pipeline (kata enn3).
